requirements.txt
asgiref==3.6.0
azure-core==1.26.4
azure-storage-blob==12.16.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.1.9
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-js-asset==2.0.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
mssql-django==1.2
Pillow==9.5.0
pycparser==2.21
pyodbc==4.0.39
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
